Added reference headers and moved some code.
