+++ OndselSolver/EulerParameters.ref.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace MbD {
+    template<typename T>
+    class EulerParameters;
+}
