+++ OndselSolver/FullColumn.ref.h
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace MbD {
+    template<typename T>
+    class FullColumn;
+
+    using FColDsptr = std::shared_ptr<FullColumn<double>>;
+
+    template<typename T>
+    using FColsptr = std::shared_ptr<FullColumn<T>>;
+}
