+++ OndselSolver/FullMatrix.h
@@ -11,38 +11,14 @@
 #include "corecrt_math_defines.h"
 #include <memory>
 
-#include "FullRow.h"
+#include "FullMatrix.ref.h"
+#include "FullColumn.ref.h"
+#include "FullRow.ref.h"
+#include "DiagonalMatrix.ref.h"
+#include "EulerParameters.ref.h"
 #include "RowTypeMatrix.h"
-#include "FullColumn.h"
 
 namespace MbD {
-    template<typename T>
-    class FullMatrix;
-
-	using FMatDsptr = std::shared_ptr<FullMatrix<double>>;
-
-	template<typename T>
-	using FMatsptr = std::shared_ptr<FullMatrix<T>>;
-
-	template<typename T>
-	class FullColumn;
-	using FColDsptr = std::shared_ptr<FullColumn<double>>;
-
-	template<typename T>
-	class FullRow;
-
-    template<typename T>
-    class RowTypeMatrix;
-
-	template<typename T>
-	class EulerParameters;
-
-	template<typename T>
-	class DiagonalMatrix;
-
-	using FMatFColDsptr = std::shared_ptr<FullMatrix<FColDsptr>>;
-	using FMatFMatDsptr = std::shared_ptr<FullMatrix<FMatDsptr>>;
-	using FColFMatDsptr = std::shared_ptr<FullColumn<FMatDsptr>>;
 
 	template<typename T>
 	class FullMatrix : public RowTypeMatrix<FRowsptr<T>>
@@ -118,210 +94,5 @@ namespace MbD {
 
 		std::ostream& printOn(std::ostream& s) const override;
 	};
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatex(T the)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, 1.0);
-		row0->atiput(1, 0.0);
-		row0->atiput(2, 0.0);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, 0.0);
-		row1->atiput(1, cthe);
-		row1->atiput(2, -sthe);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, 0.0);
-		row2->atiput(1, sthe);
-		row2->atiput(2, cthe);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatey(T the)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, cthe);
-		row0->atiput(1, 0.0);
-		row0->atiput(2, sthe);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, 0.0);
-		row1->atiput(1, 1.0);
-		row1->atiput(2, 0.0);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, -sthe);
-		row2->atiput(1, 0.0);
-		row2->atiput(2, cthe);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatez(T the)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, cthe);
-		row0->atiput(1, -sthe);
-		row0->atiput(2, 0.0);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, sthe);
-		row1->atiput(1, cthe);
-		row1->atiput(2, 0.0);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, 0.0);
-		row2->atiput(1, 0.0);
-		row2->atiput(2, 1.0);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatexrotDot(T the, T thedot)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto sthedot = cthe * thedot;
-		auto cthedot = -sthe * thedot;
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, 0.0);
-		row0->atiput(1, 0.0);
-		row0->atiput(2, 0.0);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, 0.0);
-		row1->atiput(1, cthedot);
-		row1->atiput(2, -sthedot);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, 0.0);
-		row2->atiput(1, sthedot);
-		row2->atiput(2, cthedot);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotateyrotDot(T the, T thedot)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto sthedot = cthe * thedot;
-		auto cthedot = -sthe * thedot;
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, cthedot);
-		row0->atiput(1, 0.0);
-		row0->atiput(2, sthedot);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, 0.0);
-		row1->atiput(1, 0.0);
-		row1->atiput(2, 0.0);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, -sthedot);
-		row2->atiput(1, 0.0);
-		row2->atiput(2, cthedot);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatezrotDot(T the, T thedot)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto sthedot = cthe * thedot;
-		auto cthedot = -sthe * thedot;
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, cthedot);
-		row0->atiput(1, -sthedot);
-		row0->atiput(2, 0.0);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, sthedot);
-		row1->atiput(1, cthedot);
-		row1->atiput(2, 0.0);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, 0.0);
-		row2->atiput(1, 0.0);
-		row2->atiput(2, 0.0);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatexrotDotrotDDot(T the, T thedot, T theddot)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto sthedot = cthe * thedot;
-		auto cthedot = -sthe * thedot;
-		auto stheddot = cthedot * thedot + (cthe * theddot);
-		auto ctheddot = -(sthedot * thedot) - (sthe * theddot);
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, 0.0);
-		row0->atiput(1, 0.0);
-		row0->atiput(2, 0.0);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, 0.0);
-		row1->atiput(1, ctheddot);
-		row1->atiput(2, -stheddot);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, 0.0);
-		row2->atiput(1, stheddot);
-		row2->atiput(2, ctheddot);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotateyrotDotrotDDot(T the, T thedot, T theddot)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto sthedot = cthe * thedot;
-		auto cthedot = -sthe * thedot;
-		auto stheddot = cthedot * thedot + (cthe * theddot);
-		auto ctheddot = -(sthedot * thedot) - (sthe * theddot);
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, ctheddot);
-		row0->atiput(1, 0.0);
-		row0->atiput(2, stheddot);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, 0.0);
-		row1->atiput(1, 0.0);
-		row1->atiput(2, 0.0);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, -stheddot);
-		row2->atiput(1, 0.0);
-		row2->atiput(2, ctheddot);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::rotatezrotDotrotDDot(T the, T thedot, T theddot)
-	{
-		auto sthe = std::sin(the);
-		auto cthe = std::cos(the);
-		auto sthedot = cthe * thedot;
-		auto cthedot = -sthe * thedot;
-		auto stheddot = cthedot * thedot + (cthe * theddot);
-		auto ctheddot = -(sthedot * thedot) - (sthe * theddot);
-		auto rotMat = std::make_shared<FullMatrix<T>>(3, 3);
-		auto row0 = rotMat->at(0);
-		row0->atiput(0, ctheddot);
-		row0->atiput(1, -stheddot);
-		row0->atiput(2, 0.0);
-		auto row1 = rotMat->at(1);
-		row1->atiput(0, stheddot);
-		row1->atiput(1, ctheddot);
-		row1->atiput(2, 0.0);
-		auto row2 = rotMat->at(2);
-		row2->atiput(0, 0.0);
-		row2->atiput(1, 0.0);
-		row2->atiput(2, 0.0);
-		return rotMat;
-	}
-	template<typename T>
-	inline FMatsptr<T> FullMatrix<T>::identitysptr(int n)
-	{
-		auto mat = std::make_shared<FullMatrix<T>>(n, n);
-		mat->identity();
-		return mat;
-	}
 }
